neural_net_test/requirements.txt
tensorflow-macos>=2.15
tensorflow-metal>=1.1
pandas
numpy
scikit-learn
matplotlib
